Vulkan: Use 64-bit counters
Some upcoming counters don't fit in 32 bits.

Bug: angleproject:5881
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3679488
Commit-Queue: Charlie Lao <cclao@google.com>
Auto-Submit: Shahbaz Youssefi <syoussefi@chromium.org>

+++ src/common/angleutils.h
@@ -98,7 +98,7 @@ struct PerfMonitorCounter
     ~PerfMonitorCounter();
 
     std::string name;
-    uint32_t value;
+    uint64_t value;
 };
 using PerfMonitorCounters = std::vector<PerfMonitorCounter>;
 
@@ -127,7 +127,7 @@ struct PerfMonitorTriplet
 {
     uint32_t group;
     uint32_t counter;
-    uint32_t value;
+    uint64_t value;
 };
 
 #define ANGLE_VK_PERF_COUNTERS_X(FN)               \
@@ -183,7 +183,7 @@ struct PerfMonitorTriplet
     FN(allocateNewBufferBlockCalls)                \
     FN(dynamicBufferAllocations)
 
-#define ANGLE_DECLARE_PERF_COUNTER(COUNTER) uint32_t COUNTER;
+#define ANGLE_DECLARE_PERF_COUNTER(COUNTER) uint64_t COUNTER;
 
 struct VulkanPerfCounters
 {
