+++ src/libANGLE/OverlayWidgets.cpp
@@ -140,7 +140,7 @@ void GetTextString(const std::string &src, uint8_t textOut[kMaxTextLength])
     }
 }
 
-void GetGraphValues(const std::vector<size_t> srcValues,
+void GetGraphValues(const std::vector<uint64_t> srcValues,
                     size_t startIndex,
                     float scale,
                     uint32_t valuesOut[kMaxGraphDataSize])
@@ -154,13 +154,13 @@ void GetGraphValues(const std::vector<size_t> srcValues,
     }
 }
 
-std::vector<size_t> CreateHistogram(const std::vector<size_t> values)
+std::vector<uint64_t> CreateHistogram(const std::vector<uint64_t> values)
 {
-    std::vector<size_t> histogram(values.size(), 0);
+    std::vector<uint64_t> histogram(values.size(), 0);
 
-    for (size_t rank : values)
+    for (uint64_t rank : values)
     {
-        ++histogram[rank];
+        ++histogram[static_cast<size_t>(rank)];
     }
 
     return histogram;
@@ -202,7 +202,7 @@ class AppendWidgetDataHelper
                                  TextWidgetData *textWidget,
                                  OverlayWidgetCounts *widgetCounts);
 
-    using FormatGraphTitleFunc = std::function<std::string(size_t curValue, size_t maxValue)>;
+    using FormatGraphTitleFunc = std::function<std::string(uint64_t curValue, uint64_t maxValue)>;
     static void AppendRunningGraphCommon(const overlay::Widget *widget,
                                          const gl::Extents &imageExtent,
                                          TextWidgetData *textWidget,
@@ -221,7 +221,7 @@ class AppendWidgetDataHelper
 
     static void AppendGraphCommon(const overlay::Widget *widget,
                                   const gl::Extents &imageExtent,
-                                  const std::vector<size_t> runningValues,
+                                  const std::vector<uint64_t> runningValues,
                                   size_t startIndex,
                                   float scale,
                                   GraphWidgetData *graphWidget,
@@ -244,7 +244,7 @@ void AppendWidgetDataHelper::AppendTextCommon(const overlay::Widget *widget,
 
 void AppendWidgetDataHelper::AppendGraphCommon(const overlay::Widget *widget,
                                                const gl::Extents &imageExtent,
-                                               const std::vector<size_t> runningValues,
+                                               const std::vector<uint64_t> runningValues,
                                                size_t startIndex,
                                                float scale,
                                                GraphWidgetData *graphWidget,
@@ -279,9 +279,9 @@ void AppendWidgetDataHelper::AppendRunningGraphCommon(
     const overlay::RunningGraph *matchToGraph =
         static_cast<const overlay::RunningGraph *>(matchToWidget);
 
-    const size_t maxValue =
+    const uint64_t maxValue =
         *std::max_element(graph->runningValues.begin(), graph->runningValues.end());
-    const size_t maxValueInMatchToGraph =
+    const uint64_t maxValueInMatchToGraph =
         *std::max_element(matchToGraph->runningValues.begin(), matchToGraph->runningValues.end());
     const int32_t graphHeight = std::abs(widget->coords[3] - widget->coords[1]);
     const float graphScale    = static_cast<float>(graphHeight) / maxValueInMatchToGraph;
@@ -289,7 +289,7 @@ void AppendWidgetDataHelper::AppendRunningGraphCommon(
     const size_t graphSize  = graph->runningValues.size();
     const size_t currentIdx = graph->lastValueIndex - 1;
 
-    const size_t curValue = graph->runningValues[(graphSize + currentIdx) % graphSize];
+    const uint64_t curValue = graph->runningValues[(graphSize + currentIdx) % graphSize];
 
     AppendGraphCommon(widget, imageExtent, graph->runningValues, graph->lastValueIndex + 1,
                       graphScale, graphWidget, widgetCounts);
@@ -313,13 +313,13 @@ void AppendWidgetDataHelper::AppendRunningHistogramCommon(const overlay::Widget 
     const overlay::RunningHistogram *runningHistogram =
         static_cast<const overlay::RunningHistogram *>(widget);
 
-    std::vector<size_t> histogram = CreateHistogram(runningHistogram->runningValues);
-    auto peakRangeIt              = std::max_element(histogram.rbegin(), histogram.rend());
-    const size_t peakRangeValue   = *peakRangeIt;
-    const int32_t graphHeight     = std::abs(widget->coords[3] - widget->coords[1]);
-    const float graphScale        = static_cast<float>(graphHeight) / peakRangeValue;
-    auto maxValueIter =
-        std::find_if(histogram.rbegin(), histogram.rend(), [](size_t value) { return value != 0; });
+    std::vector<uint64_t> histogram = CreateHistogram(runningHistogram->runningValues);
+    auto peakRangeIt                = std::max_element(histogram.rbegin(), histogram.rend());
+    const uint64_t peakRangeValue   = *peakRangeIt;
+    const int32_t graphHeight       = std::abs(widget->coords[3] - widget->coords[1]);
+    const float graphScale          = static_cast<float>(graphHeight) / peakRangeValue;
+    auto maxValueIter               = std::find_if(histogram.rbegin(), histogram.rend(),
+                                                   [](uint64_t value) { return value != 0; });
 
     AppendGraphCommon(widget, imageExtent, histogram, 0, graphScale, graphWidget, widgetCounts);
 
@@ -383,7 +383,7 @@ void AppendWidgetDataHelper::AppendVulkanRenderPassCount(const overlay::Widget *
                                                          GraphWidgetData *graphWidget,
                                                          OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "RenderPass Count: " << maxValue;
         return text.str();
@@ -432,7 +432,7 @@ void AppendWidgetDataHelper::AppendVulkanWriteDescriptorSetCount(const overlay::
                                                                  GraphWidgetData *graphWidget,
                                                                  OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "WriteDescriptorSet Count: " << maxValue;
         return text.str();
@@ -447,7 +447,7 @@ void AppendWidgetDataHelper::AppendVulkanDescriptorSetAllocations(const overlay:
                                                                   GraphWidgetData *graphWidget,
                                                                   OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Descriptor Set Allocations: " << maxValue;
         return text.str();
@@ -462,7 +462,7 @@ void AppendWidgetDataHelper::AppendVulkanShaderResourceDSHitRate(const overlay::
                                                                  GraphWidgetData *graphWidget,
                                                                  OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Shader Resource DS Hit Rate (Max: " << maxValue << "%)";
         return text.str();
@@ -477,7 +477,7 @@ void AppendWidgetDataHelper::AppendVulkanDynamicBufferAllocations(const overlay:
                                                                   GraphWidgetData *graphWidget,
                                                                   OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "DynamicBuffer Allocations (Max: " << maxValue << ")";
         return text.str();
@@ -493,7 +493,7 @@ void AppendWidgetDataHelper::AppendVulkanTextureDescriptorCacheSize(
     GraphWidgetData *graphWidget,
     OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Total Texture Descriptor Cache Size: " << curValue;
         return text.str();
@@ -509,7 +509,7 @@ void AppendWidgetDataHelper::AppendVulkanUniformDescriptorCacheSize(
     GraphWidgetData *graphWidget,
     OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Total Uniform Descriptor Cache Size: " << curValue;
         return text.str();
@@ -524,7 +524,7 @@ void AppendWidgetDataHelper::AppendVulkanDescriptorCacheSize(const overlay::Widg
                                                              GraphWidgetData *graphWidget,
                                                              OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Total Descriptor Cache Size: " << curValue;
         return text.str();
@@ -553,7 +553,7 @@ void AppendWidgetDataHelper::AppendVulkanAttemptedSubmissions(const overlay::Wid
                                                               GraphWidgetData *graphWidget,
                                                               OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Attempted submissions (peak): " << maxValue;
         return text.str();
@@ -568,7 +568,7 @@ void AppendWidgetDataHelper::AppendVulkanActualSubmissions(const overlay::Widget
                                                            GraphWidgetData *graphWidget,
                                                            OverlayWidgetCounts *widgetCounts)
 {
-    auto format = [](size_t curValue, size_t maxValue) {
+    auto format = [](uint64_t curValue, uint64_t maxValue) {
         std::ostringstream text;
         text << "Actual submissions (peak): " << maxValue;
         return text.str();
